+++ libs/base_num/src/int/wrapper/namespace.rs
@@ -5,7 +5,7 @@
 
 #[doc = crate::TAG_NUM!()]
 #[doc = crate::TAG_NAMESPACE!()]
-/// Provides comprehensive integer operations on `T`, most of them *const*.
+/// Provides comprehensive integer operations on `T`, all of them *const*.
 ///
 /// It's implemented for:
 /// - all the integer primitives: `i8`, …, `i128`, `u8`, …, `u128`.